feat(payment): PAYPAL-3603 updated braintree analytic tracker with fastlane support
