+++ packages/braintree-utils/src/braintree.ts
@@ -714,7 +714,7 @@ export interface BraintreeConnect {
     ConnectCardComponent: (
         options: BraintreeConnectCardComponentOptions,
     ) => BraintreeConnectCardComponent;
-    events: BraintreeConnectEvents;
+    events: BraintreeFastlaneEvents;
 }
 
 export interface BraintreeConnectWindow extends Window {
@@ -852,45 +852,6 @@ export interface BraintreeConnectTokenizeDetails {
 export interface BraintreeConnectTokenizeOptions {
     billingAddress?: BraintreeConnectAddress;
     shippingAddress?: BraintreeConnectAddress;
-}
-
-export interface BraintreeConnectEvents {
-    apmSelected: (options: BraintreeConnectApmSelectedEventOptions) => void;
-    emailSubmitted: (options: BraintreeConnectEmailEnteredEventOptions) => void;
-    orderPlaced: (options: BraintreeConnectOrderPlacedEventOptions) => void;
-}
-
-export interface BraintreeConnectEventCommonOptions {
-    context_type: 'cs_id';
-    context_id: string; // checkout session id
-    page_type: 'checkout_page';
-    page_name: string; // title of the checkout initiation page
-    partner_name: 'bigc';
-    user_type: 'store_member' | 'store_guest'; // type of the user on the merchant site
-    store_id: string;
-    merchant_name: string;
-    experiment: string; // stringify JSON object "[{ treatment_group: 'test' | 'control' }]"
-}
-
-export interface BraintreeConnectApmSelectedEventOptions
-    extends BraintreeConnectEventCommonOptions {
-    apm_shown: '0' | '1'; // alternate payment shown on the checkout page
-    apm_list: string; // list of alternate payment shown on checkout page
-    apm_selected: string; // alternate payment method selected / methodId
-    apm_location: 'pre-email section' | 'payment section'; // placement of APM, whether it be above the email entry or in the radio buttons
-}
-
-export interface BraintreeConnectEmailEnteredEventOptions
-    extends BraintreeConnectEventCommonOptions {
-    user_email_saved: boolean; // shows whether checkout was loaded with or without a saved email
-    apm_shown: '0' | '1'; // alternate payment shown on the checkout page
-    apm_list: string; // list of alternate payment shown on checkout page 'applepay,googlepay,paypal'
-}
-
-export interface BraintreeConnectOrderPlacedEventOptions
-    extends BraintreeConnectEventCommonOptions {
-    selected_payment_method: string;
-    currency_code: string;
 }
 
 export interface Card {
