+++ packages/braintree-utils/src/utils/index.ts
@@ -1,3 +1,4 @@
 export { default as getValidBraintreeFastlaneStyles } from './get-valid-braintree-fastlane-styles';
 export { default as isBraintreeConnectWindow } from './is-braintree-connect-window';
+export { default as isBraintreeFastlaneWindow } from './is-braintree-fastlane-window';
 export { default as isBraintreeAcceleratedCheckoutCustomer } from './is-braintree-accelerated-checkout-customer';
