+++ packages/braintree-utils/src/utils/is-braintree-fastlane-window.spec.ts
@@ -1,0 +1,15 @@
+import { BraintreeFastlaneWindow } from '../braintree';
+
+import isBraintreeFastlaneWindow from './is-braintree-fastlane-window';
+
+describe('isBraintreeFastlaneWindow', () => {
+    it('window has braintreeFastlane option', () => {
+        expect(
+            isBraintreeFastlaneWindow({ braintreeFastlane: {} } as BraintreeFastlaneWindow),
+        ).toBe(true);
+    });
+
+    it('window does not have braintreeFastlane option', () => {
+        expect(isBraintreeFastlaneWindow({} as Window)).toBe(false);
+    });
+});
