+++ packages/braintree-utils/src/utils/is-braintree-fastlane-window.ts
@@ -1,0 +1,7 @@
+import { BraintreeFastlaneWindow } from '../braintree';
+
+export default function isBraintreeFastlaneWindow(
+    window: Window,
+): window is BraintreeFastlaneWindow {
+    return Boolean(window.hasOwnProperty('braintreeFastlane'));
+}
