+++ packages/core/src/analytics/braintree-connect-tracker/braintree-analytic-tracker-service.ts
@@ -1,0 +1,6 @@
+export default interface BraintreeAnalyticTrackerService {
+    customerPaymentMethodExecuted(): void;
+    paymentComplete(): void;
+    selectedPaymentMethod(methodId: string): void;
+    walletButtonClick(methodId: string): void;
+}
